# Server/openapi.yaml
openapi: 3.0.0
info:
  title: TestVar - Flashcards API
  description: "A revolutionary REST API for flashcards"
  version: 1.0.0
servers:
  - url: http://localhost:7222/api
    description: Development server
paths:
...
  /sets:
    get:
      summary: "Get all flashcard sets"
      tags:
        - "FlashCard Sets"
      responses:
        "200":
          description: "A list of flashcard sets"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/FlashcardSet"
    post:
      summary: "Create a new flashcard set"
      description: "There can only be 20 new flashcard sets created per day"
      tags:
        - "FlashCard Sets"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardSet"
      responses:
        "201":
          description: "The created flashcard set"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardSet"
        "429":
          description: "You have reached the maximum number of flashcard sets allowed today"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /sets/{setId}:
    parameters:
      - name: setId
        in: path
        required: true
        description: "The ID of the flashcard set"
        schema:
          type: "string"
    get:
      summary: "Get a flashcard set by ID"
      tags:
        - "FlashCard Sets"
      responses:
        "200":
          description: "The flashcard set"
          content:
            application/json:
              schema:
                type: "object"
                allOf:
                  - $ref: "#/components/schemas/FlashcardSet"
                  - type: "object"
                    properties:
                      comments:
                        type: "array"
                        description: "Comments on the set"
                        items:
                          type: "object"
                          properties:
                            comment:
                              type: "string"
                              example: "I love this set!"
                            author:
                              $ref: "#/components/schemas/User"
        "404":
          description: "The flashcard set was not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    put:
      summary: "Update a flashcard set by ID"
      tags:
        - "FlashCard Sets"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardSet"
      responses:
        "200":
          description: "The updated flashcard set"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardSet"
        "404":
          description: "The flashcard set was not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    delete:
      summary: "Delete a flashcard set by ID"
      tags:
        - "FlashCard Sets"
      responses:
        "204":
          description: "The flashcard set was deleted"
        "404":
          description: "The flashcard set was not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
  /set/{setID}/comment:
    parameters:
      - name: setID
        in: path
        required: true
        description: "The ID of the flashcard set"
        schema:
          type: "string"
    post:
      summary: "Comment on a flashcard set, by the current user"
      tags:
        - "FlashCard Sets"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                comment:
                  type: "string"
                  example: "I love this set!"
      responses:
        "201":
          description: "The created comment"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "404":
          description: "The flashcard set was not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /sets/{setId}/cards:
    parameters:
      - name: setId
        in: path
        required: true
        description: "The ID of the flashcard set"
        schema:
          type: "string"
    get:
      summary: "Get all flashcards in a set"
      tags:
        - "FlashCard Sets"
      parameters:
        - name: shuffle
          in: query
          required: false
          description: "Get the flashcards in a random order"
          schema:
            type: "boolean"
      responses:
        "200":
          description: "A list of flashcards"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/FlashCard"
        "404":
          description: "The flashcard set was not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
  /users/{userID}/sets:
    get:
      summary: "Get all flashcard sets created by a user"
      tags:
        - "Users"
        - "FlashCard Sets"
      parameters:
        - name: userID
          in: path
          required: true
          description: "The ID of the user"
          schema:
            type: "string"
      responses:
        "200":
          description: "A list of flashcard sets"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/FlashcardSet"
        "404":
          description: "The user was not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
components:
  schemas:
...
    FlashCard:
      type: "object"
      properties:
        question:
          type: "string"
          example: "What is the capital of France?"
        answer:
          type: "string"
          example: "Paris"
        difficulty:
          $ref: "#/components/schemas/Difficulty"
          nullable: true
    FlashcardSet:
      type: "object"
      properties:
        id:
          type: "number"
          example: 1
        name:
          type: "string"
          example: "European Capitals"
        cards:
          type: "array"
          items:
            $ref: "#/components/schemas/FlashCard"
        created_at:
          type: "string"
          format: "date-time"
          example: "2024-10-14T12:00:00Z"
        updated_at:
          type: "string"
          format: "date-time"
          example: "2024-10-15T12:00:00Z"
    ProblemDetails:
      type: "object"
      properties:
        title:
          type: "string"
          nullable: false
          description: "Summary of the problem"
        status:
          type: "number"
          nullable: false
          description: "HTTP status code"
        detail:
          type: "string"
          nullable: true
          description: "Additional problem details"
    Difficulty:
      type: "string"
      enum:
        - "easy"
        - "medium"
        - "hard"
    User:
      type: "object"
      properties:
        id:
          type: "number"
          example: 1
        username:
          type: "string"
          example: "johndoe"
        admin:
          type: "boolean"
          example: false
          default: false
          nullable: true
          description: "The set creation limit should only apply to non-admin users"
    Comment:
      type: "object"
      properties:
        comment:
          type: "string"
          example: "I love this set!"
        set:
          $ref: "#/components/schemas/FlashcardSet"
        author:
          $ref: "#/components/schemas/User"
...